requirements.txt
python-dotenv>=1.0.0, <2.0.0
redis==5.2.0
pydantic==2.9.2
crewai==0.102.0
tc-temporal-backend==1.1.0
